feat(audio): split the audio interface from the backend library
This paves the way to support multiple audio backends
